Patient phone is now long int

+++ core/src/main/java/org/teamone/core/Patient.java
@@ -12,7 +12,7 @@ public class Patient extends Person{
     private String gender;
     private String insurance;
     private int age = 0;
-    private int phone = 0;
+    private long phone = 0;
     private String healthConditions;
     private String labReports;
     private String alertDateAndTime;
@@ -83,7 +83,7 @@ public class Patient extends Person{
         this.age = age;
     }
 
-    public int getPhone() {
+    public long getPhone() {
         return phone;
     }
